requirement.txt
transformers
torch
pillow
streamlit